webapi/webapi-shadowdom-w3c-tests/shadowdom/w3c/elements-and-dom-objects/csshostrule-interface/csshostrule-attributes/CSSHostRule.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Wang, Hongjuan<hongjuanx.wang@intel.com>

-->

<html>
  <head>
    <meta charset='utf-8'>
    <title>ShadowDom Test: CSSHostRule</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/shadow/index.html#api-shadow-root-get-element-by-id">
    <script src="../../../../../resources/testharness.js"></script>
    <script src="../../../../../resources/testharnessreport.js"></script>
    <script src="../../../testcommon.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(unit(function (ctx){
        var d = newRenderedHTMLDocument(ctx);
        var host = d.createElement('div');
          d.body.appendChild(host);
          createSR(host);
        var style = d.createElement('style');
          style.innerHTML = '' +
			    'span { color: blue; }' +
			    '@host {' +
			    'div { color: red; }' +
			    '}';
	      d.head.appendChild(style);
	      CSSHostRule = d.styleSheets[0].cssRules[1];
        assert_not_equals(CSSHostRule.cssRules, null,	"the value of CSSHostRule.cssRules is not null");
      }), "Check if on getting, the attribute must return the list of all CSS rules.");

      test(unit(function (ctx){
        var d = newRenderedHTMLDocument(ctx);
        var host = d.createElement('div');
          d.body.appendChild(host);
          createSR(host);
        var style = d.createElement('style');
          style.innerHTML = '' +
			    'span { color: blue; }' +
			    '@host {' +
			    'div { color: red; }' +
			    '}';
	      d.head.appendChild(style);
	      CSSHostRule = d.styleSheets[0].cssRules[1];
	      CSSHostRule.insertRule("div {width: 200px; }", 1);
        assert_true(true,	"the method of CSSHostRule.cssRules is valid");
      }), "Check if the method must insert a CSS rule rule into the CSS rule list returned by cssRules at index when invoked.");

      test(unit(function (ctx){
        var d = newRenderedHTMLDocument(ctx);
        var host = d.createElement('div');
          d.body.appendChild(host);
          createSR(host);
        var style = d.createElement('style');
          style.innerHTML = '' +
			    'span { color: blue; }' +
			    '@host {' +
			    'div { color: red; }' +
			    '}';
	      d.head.appendChild(style);
	      CSSHostRule = d.styleSheets[0].cssRules[1];
	      CSSHostRule.deleteRule(0);
        assert_true(true,	"the method of CSSHostRule.deleteRule is valid");
      }), "Check if the method must remove a CSS rule rule from the CSS rule list returned by cssRules at index when invoked.");
    </script>
  </body>
</html>
